Add c vpdfonly max1mb

+++ RecruitmentTracking/Controllers/AdminController.cs
@@ -89,6 +89,7 @@ public class AdminController : Controller
         return View(listJob);
     }
 
+    
     [HttpPost]
     public async Task<IActionResult> CloseTheJob(int id)
     {
@@ -113,11 +114,19 @@ public class AdminController : Controller
         return Redirect("/JobClosed");
     }
 
+
+    // Add Feature, if candidate apply job > 0, job can't be closed
     [HttpPost]
     public async Task<IActionResult> DeleteJob(int id)
     {
-        Job objJob = _db.Jobs!.Find(id)!;
+        if(_db.CandidateJobs!.Where(cj => cj.JobId == id).Count() > 0)
+        {
+            TempData["warning"] = "Job can't be closed because there are candidates who have applied for this job.";
+            //return Redirect("/Admin");
+            return Redirect("/JobClosed");
+        }
 
+        Job objJob = _db.Jobs!.Find(id)!;
         _db.Jobs.Remove(objJob);
         await _db.SaveChangesAsync();
